código refatorado

+++ src/App.js
@@ -1,26 +1,35 @@
 import express from 'express';
 import morgan from 'morgan';
+import * as Sentry from '@sentry/node';
 
+import sentryConfig from './config/sentry';
 import allowCors from './middlewares/cors';
-import deleteTempFiles from './middlewares/deleteTempFiles';
+import deleteTempFiles from './utils/deleteTempFiles';
 
-import routes from './routes';
+import routes from './api/routes';
 
 class App {
   constructor() {
-    // this.nodeEnv = process.env.NODE_ENV;
+    this.nodeEnv = process.env.NODE_ENV;
     // this.subDirectory = process.env.SUBDIRECTORY;
 
     this.server = express();
+
+    Sentry.init(sentryConfig);
 
     this.middlewares();
     this.routes();
   }
 
   middlewares() {
-    deleteTempFiles();
+    this.server.use(Sentry.Handlers.requestHandler());
+
+    if (this.nodeEnv === 'prod') deleteTempFiles();
+
     this.server.use(express.json());
     this.server.use(allowCors);
+    this.server.use(express.static('./public'));
+    this.server.use('/public', express.static('./public'));
 
     this.server.use(
       morgan(`[:date] - :method [:status] :url - :response-time ms`)
@@ -29,6 +38,8 @@ class App {
 
   routes() {
     this.server.use('/api', routes);
+    this.server.get('/favicon.ico', (req, res) => res.status(204));
+    this.server.use(Sentry.Handlers.errorHandler());
     // this.server.use(`${this.subDirectory}/v2`, apiRoutesV2);
   }
 }
